// src/shared/api/kinopoisk/models.ts
export type genresList = { genre: string}[];
export type countriesList = { country: string }[];

export interface CatalogItem {
  nameRu: string;
  posterUrlPreview: string;
  posterUrl: string;
  ratingKinopoisk: number;
  kinopoiskId: number;
  year: number;
  genres: genresList;
  countries: countriesList;
}

export type CatalogList = CatalogItem[];

export interface GetCatalogListResponse {
  total: number,
  totalPages: number,
  items: CatalogList,
}

export enum catalogOrderTypes {
  RATING = 'RATING',
  NUM_VOTE = 'NUM_VOTE',
  YEAR = 'YEAR'
}
export enum catalogTypeTypes{
  FILM = 'FILM',
  TV_SHOW = 'TV_SHOW',
  TV_SERIES = 'TV_SERIES',
  MINI_SERIES = 'MINI_SERIES',
  ALL = 'ALL'
}
// ...
export interface CatalogParams {
  countries?: number[];
  genres?: number[];
  order?: catalogOrderTypes;
  type?: catalogTypeTypes;
  ratingFrom?: number;
  ratingTo?: number;
  yearFrom?: number;
  yearTo?: number;
  keyword?: string;
  page?: number;
}

export interface Film {
  nameRu?: string;
  nameEn?: string;
  nameOriginal?: string;
  posterUrl?: string;
  description?: string;
  ratingAgeLimits?: string;
  kinopoiskId?: number;
  coverUrl?: string | null;
  logoUrl?: string | null;
  ratingKinopoisk?: number;
  webUrl?: string;
  year?: number;
  filmLength?: number;
  slogan?: string;
  shortDescription?: string;
  type?: catalogTypeTypes;
  genres?: genresList;
  countries?: countriesList;
  startYear?: null;
  endYear?: null;
  completed?: false;
}

export const FilmType: Record<catalogTypeTypes, string> = {
  [catalogTypeTypes.FILM]: 'Фильмы',
  [catalogTypeTypes.TV_SHOW]: 'TV-шоу',
  [catalogTypeTypes.TV_SERIES]: 'TV-сериал',
  [catalogTypeTypes.MINI_SERIES]: 'Мини-сериал',
  [catalogTypeTypes.ALL]: 'Все',
};

export interface FilmBudgetItem {
  type: string;
  amount: number;
  currencyCode: string;
  name: string;
  symbol: string;
}
export interface FilmBudget {
  total: number;
  totalPages: number;
  items: FilmBudgetItem[];
}

export interface FilmImageItem {
  imageUrl: string;
  previewUrl: string;
}

export interface FilmImages {
  total: number;
  totalPages: number;
  items: FilmImageItem[];
}

export enum FilmFactItemType {
  FACT = 'FACT',
  BLOOPER = 'BLOOPER'
}

export interface FilmFactItem {
  text: string;
  type: FilmFactItemType;
  spoiler: boolean;
}

export interface FilmFacts {
  total: number;
  items: FilmFactItem[];
}

export enum TeamProfession {
  DIRECTOR = 'Режиссер',
  DESIGN = 'Художник',
  EDITOR = 'Монтажер',
  WRITER = 'Сценарист',
  VOICE_DIRECTOR = 'Режиссер дубляжа',
  COMPOSER = 'Композитор',
  OPERATOR = 'Оператор',
  PRODUCER = 'Продюсер',
  ACTOR = 'Актер',
  HIMSELF = 'Играет самого себя',
  HERSELF = 'Играет саму себя',
  HRONO_TITR_MALE = 'Мужской голос за кадром',
  HRONO_TITR_FEMALE = 'Женский голос за кадром',
  VOICE_MALE = 'Актер озвучки',
  VOICE_FEMALE = 'Акриса озвучки',
}

export interface FilmTeamItem {
  staffId: number;
  nameEn: string;
  nameRu: string;
  description: null | string;
  posterUrl: string;
  professionText: string;
  professionKey: TeamProfession;
}
export type FilmTeam = FilmTeamItem[];

export interface Spouse {
  personId: number;
  name: string;
  divorced: boolean;
  divorcedReason: string;
  sex: string;
  children: number;
  webUrl: string;
  relation: string;
}
export interface PersonFilmTypes {
  filmId: number;
  nameRu: string;
  nameEn: string;
  rating: number | null;
  general: boolean;
  description: string;
  professionKey: string;
}
export interface Person {
  personId?: number;
  webUrl?: string ;
  nameRu?: string | null;
  nameEn?: string | null;
  sex?: string;
  posterUrl?: string;
  growth?: number;
  birthday?: string;
  death?: string | null;
  age?: number;
  birthplace?: string;
  deathplace?: string | null;
  spouses?: Spouse[];
  hasAwards?: number;
  profession?: string;
  facts?: string[];
  films?: PersonFilmTypes[];
}